refactor: Internal package
Move handler code to "internal" package.
